move: chat 관련 훅 파일 chat 폴더 안으로 이동 (#47)
